+++ src/widgets/classTree/classTree.tsx
@@ -28,7 +28,7 @@ import {
 import { WorkspaceContext, useWorkspace } from '../../workspace/workspaceContext';
 
 import { TreeNode } from './treeModel';
-import { Forest } from './leaf';
+import { ClassTreeContext, Forest } from './leaf';
 
 /**
  * Props for {@link ClassTree} component.
@@ -59,6 +59,12 @@ export interface ClassTreeProps {
      * @default 2
      */
     minSearchTermLength?: number;
+    /**
+     * Whether tree items can be dragged out e.g. onto the canvas.
+     *
+     * @default true
+     */
+    draggableItems?: boolean;
 }
 
 /**
@@ -160,7 +166,7 @@ class ClassTreeInner extends React.Component<ClassTreeInnerProps, State> {
     render() {
         const {
             className, isControlled, searchStore, normalizedTerm, minSearchTermLength,
-            workspace: {editor}, translation: t,
+            draggableItems = true, workspace: {editor}, translation: t,
         } = this.props;
         const {
             fetchedGraph, refreshingState, appliedSearchText, roots, filteredRoots, selectedNode,
@@ -202,30 +208,35 @@ class ClassTreeInner extends React.Component<ClassTreeInnerProps, State> {
                     title={t.text('search_element_types.refresh_progress.title')}
                 />
                 {fetchedGraph?.classTree ? (
-                    <Forest className={`${CLASS_NAME}__tree reactodia-scrollable`}
-                        nodes={filteredRoots}
-                        searchText={searchText}
-                        selectedNode={selectedNode}
-                        onSelect={this.onSelectNode}
-                        creatableClasses={
-                            editor.inAuthoringMode ? constructibleClasses : EMPTY_CREATABLE_TYPES
-                        }
-                        onClickCreate={this.onCreateInstance}
-                        onDragCreate={this.onDragCreate}
-                        footer={
-                            filteredRoots.length === 0 ? (
-                                <NoSearchResults className={`${CLASS_NAME}__no-results`}
-                                    hasQuery={filteredRoots !== roots}
-                                    minSearchTermLength={minSearchTermLength}
-                                    message={
-                                        roots.length === 0
-                                            ? t.text('search_element_types.no_results')
-                                            : undefined
-                                    }
-                                />
-                            ) : null
-                        }
-                    />
+                    <ClassTreeContext.Provider
+                        value={{
+                            searchText,
+                            selectedNode,
+                            onSelect: this.onSelectNode,
+                            creatableClasses: editor.inAuthoringMode
+                                ? constructibleClasses : EMPTY_CREATABLE_TYPES,
+                            onClickCreate: this.onCreateInstance,
+                            onDragCreate: this.onDragCreate,
+                            draggableItems,
+                        }}>
+                        <Forest className={`${CLASS_NAME}__tree reactodia-scrollable`}
+                            nodes={filteredRoots}
+                            root={true}
+                            footer={
+                                filteredRoots.length === 0 ? (
+                                    <NoSearchResults className={`${CLASS_NAME}__no-results`}
+                                        hasQuery={filteredRoots !== roots}
+                                        minSearchTermLength={minSearchTermLength}
+                                        message={
+                                            roots.length === 0
+                                                ? t.text('search_element_types.no_results')
+                                                : undefined
+                                        }
+                                    />
+                                ) : null
+                            }
+                        />
+                    </ClassTreeContext.Provider>
                 ) : (
                     <div className={`${CLASS_NAME}__spinner`}>
                         <HtmlSpinner width={30} height={30}